+++ out/referenceProvider/batariBasicReferenceProvider.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const referenceProviderBase_1 = require("./referenceProviderBase");
+class BatariBasicReferenceProvider extends referenceProviderBase_1.ReferenceProviderBase {
+    constructor() {
+        super("batariBasic");
+    }
+}
+exports.BatariBasicReferenceProvider = BatariBasicReferenceProvider;
+//# sourceMappingURL=batariBasicReferenceProvider.js.map