+++ out/referenceProvider/seventyEightHundredBasicReferenceProvider.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const referenceProviderBase_1 = require("./referenceProviderBase");
+class SeventyEightHundredBasicReferenceProvider extends referenceProviderBase_1.ReferenceProviderBase {
+    constructor() {
+        super("7800basic");
+    }
+}
+exports.SeventyEightHundredBasicReferenceProvider = SeventyEightHundredBasicReferenceProvider;
+//# sourceMappingURL=seventyEightHundredBasicReferenceProvider.js.map