+++ src/referenceProvider/batariBasicReferenceProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { ReferenceProviderBase } from './referenceProviderBase';
+
+export class BatariBasicReferenceProvider extends ReferenceProviderBase {
+    
+    constructor() {
+        super("batariBasic");
+    }
+}