+++ src/referenceProvider/seventyEightHundredBasicReferenceProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { ReferenceProviderBase } from './referenceProviderBase';
+
+export class SeventyEightHundredBasicReferenceProvider extends ReferenceProviderBase {
+    
+    constructor() {
+        super("7800basic");
+    }
+}