+++ CHANGELOG.md
@@ -4,8 +4,8 @@ The following enhancements and changes have been made to Atari Dev Studio:
 
 ## 0.5.5
 
-* Added document outline feature for batari Basic and 7800basic
-* Added Go to Definition feature for batari Basic and 7800basic
+* Added document outline for batari Basic and 7800basic
+* Added Go to Definition and Go to References features for batari Basic and 7800basic
 
 ## 0.5.4